backend/requirements.txt
asgiref==3.9.1
Django==5.2.6
django-cors-headers==4.8.0
djangorestframework==3.16.1
pillow==11.3.0
python-decouple==3.8
pytz==2024.2
sqlparse==0.5.3
whitenoise==6.6.0
twilio==8.10.0
gunicorn==21.2.0
cryptography==43.0.3
sendgrid==6.10.0
dj-database-url==2.3.0
psycopg2-binary==2.9.9
user-agents==2.2.0

# Testing dependencies
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.1
coverage==7.6.9
freezegun==1.5.1
responses==0.25.3

# Linting dependencies
flake8==7.1.1
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
redis==5.0.1

# MCP and AI dependencies
mcp>=1.0.0
anthropic>=0.34.0
httpx>=0.27.0
pydantic>=2.0.0
